pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "argos-fim"
version = "1.0.0"
description = "Argos — the hundred-eyed guardian. A production-quality file integrity checker."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "click>=8.1.7",
    "rich>=13.7.0",
    "groq>=0.4.2",
    "PyYAML>=6.0.1",
    "schedule>=1.2.1",
]

[project.scripts]
argos = "argos.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["argos*"]
